"当前环境信息Environment" iot模块 读取迁移到DF-K10开发板下面
