+++ main/boards/common/board.h
@@ -46,9 +46,6 @@ public:
     virtual void StartNetwork() = 0;
     virtual const char* GetNetworkStateIcon() = 0;
     virtual bool GetBatteryLevel(int &level, bool& charging, bool& discharging);
-    virtual bool GetTemperature(float *temperature);
-    virtual bool GetHumidity(float *humidity);
-    virtual bool GetLight(float *light);
     virtual std::string GetJson();
     virtual void SetPowerSaveMode(bool enabled) = 0;
 };
