+++ main/boards/common/i2c_device.h
@@ -13,11 +13,6 @@ class I2cDevice {
     void WriteReg(uint8_t reg, uint8_t value);
     uint8_t ReadReg(uint8_t reg);
     void ReadRegs(uint8_t reg, uint8_t* buffer, size_t length);
-
-    void WriteValue(uint8_t value);
-    void WriteValues(uint8_t* buffer, size_t length);
-    uint8_t ReadValue();
-    void ReadValues(uint8_t* buffer, size_t length);
 };
 
 #endif  // I2C_DEVICE_H
