+++ main/boards/df-k10/README.md
@@ -4,6 +4,12 @@
 * A：短按-打断/唤醒，长按1s-音量调大
 * B：短按-打断/唤醒，长按1s-音量调小
 
+## IoT
+* 当前环境温度
+* 当前环境湿度
+* 设置温度偏差：默认为-8
+* 设置湿度偏差：默认未+10
+
 ## 编译配置命令
 
 **配置编译目标为 ESP32S3：**
